+++ Kernel/Walls/TriangleWall.cc
@@ -153,7 +153,7 @@ void TriangleWall::write(std::ostream& os) const
 
 void TriangleWall::writeVTK (VTKContainer& vtk) const
 {
-    const int s = vtk.points.size();
+    const unsigned long s = vtk.points.size();
     vtk.points.reserve(s+3);
     for (auto v : vertex_) {
         vtk.points.push_back(v+getPosition());
